backend/requirements_fixed.txt
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
requests==2.32.3
urllib3==2.0.7
python-dotenv==1.0.1
arrow==1.3.0
